+++ src/PSRule.Rules.Azure/rules/Azure.AKS.Rule.ps1
@@ -4,7 +4,7 @@
 
 # Synopsis: AKS clusters should have minimum number of nodes for failover and updates
 Rule 'Azure.AKS.MinNodeCount' -If { ResourceType 'Microsoft.ContainerService/managedClusters' } -Tag @{ severity = 'Important'; category = 'Reliability' } {
-    Hint 'Use at least three (3) agent nodes'
+    Recommend 'Use at least three (3) agent nodes'
 
     $TargetObject.Properties.agentPoolProfiles[0].count -ge 3
 }
@@ -13,7 +13,7 @@ Rule 'Azure.AKS.MinNodeCount' -If { ResourceType 'Microsoft.ContainerService/man
 Rule 'Azure.AKS.Version' -If { ResourceType 'Microsoft.ContainerService/managedClusters' } -Tag @{ severity = 'Important'; category = 'Operations management' } {
     $minVersion = [Version]$Configuration.minAKSVersion
 
-    Hint "Upgrade Kubernetes to at least $minVersion"
+    Recommend "Upgrade Kubernetes to at least $minVersion"
 
     ([Version]$TargetObject.Properties.kubernetesVersion) -ge $minVersion
 } -Configure @{ minAKSVersion = '1.13.5' }