+++ src/PSRule.Rules.Azure/rules/Azure.SQL.Rule.ps1
@@ -4,7 +4,7 @@
 
 # Synopsis: Determine if there is an excessive number of firewall rules
 Rule 'Azure.SQL.FirewallRuleCount' -If { ResourceType 'Microsoft.Sql/servers' } -Tag @{ severity = 'Awareness'; category = 'Operations management' } {
-    Hint 'SQL Server has > 10 firewall rules, some rules may not be needed';
+    Recommend 'SQL Server has > 10 firewall rules, some rules may not be needed';
 
     $firewallRules = @($TargetObject.resources | Where-Object -FilterScript {
         $_.Type -eq 'Microsoft.Sql/servers/firewallRules'